Refactored code for login and social login
