+++ src/App.scss
@@ -1,7 +1,7 @@
-.App {
-  text-align: center;
-}
-
 * {
   font-family: "Open Sans", sans-serif;
+}
+
+.well-container {
+  margin-top: 0px;
 }
