+++ src/App.tsx
@@ -4,7 +4,8 @@ import { SpeedInsights } from "@vercel/speed-insights/react";
 import { Container } from "react-bootstrap";
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 import "./App.scss";
-import LoginForm from "./components/LoginForm";
+import ComponentsList from "./components/ComponentsList";
+import { Login } from "./pages/Login/Login";
 
 function App() {
   return (
@@ -13,7 +14,8 @@ function App() {
       <Container className="well-container">
         <BrowserRouter>
           <Routes>
-            <Route path="/" element={<LoginForm />} />
+            <Route path="/" element={<ComponentsList />} />
+            <Route path="/login" element={<Login />} />
           </Routes>
         </BrowserRouter>
       </Container>
