+++ src/anchor.scss
@@ -1,0 +1,43 @@
+/* Anchor Styling */
+@import "./colors.scss";
+
+.truncate {
+  white-space: nowrap;
+  overflow-x: hidden;
+  text-overflow: ellipsis;
+}
+
+a,
+.anchor {
+  text-decoration: none;
+  font-weight: 500;
+  transition: color 0.3s ease, text-decoration 0.3s ease;
+  font-size: 14px;
+  color: $primary;
+  cursor: pointer;
+  padding: 5px 0;
+  position: relative;
+
+  &:hover {
+    text-decoration: none;
+  }
+
+  &::after {
+    background-color: $info;
+    bottom: 0;
+    content: "";
+    height: 1px;
+    left: 0;
+    position: absolute;
+    right: 0;
+    transform: scaleX(0);
+    transition: transform 0.8s;
+    width: 100%;
+  }
+
+  &:hover::after {
+    transform: scaleX(1);
+  }
+
+  @extend .truncate;
+}
