+++ src/colors.scss
@@ -1,10 +1,10 @@
-$primary: #007bff;
+$primary: #002851;
 $secondary: #6c757d;
 $warning: #856404;
 $success: #155724;
 $info: #0c5460;
 $danger: #721c24;
-$content-color: #1c4980;
+$content-color: #000000;
 
 // for status label background
 $primary-bg: #e7f1ff;
