+++ src/modules/SocialLogin/SocialLogin.types.ts
@@ -1,0 +1,11 @@
+export interface ISocialLoginProps {
+  isGoogleAvailable?: boolean;
+  isAppleAvailable?: boolean;
+  isAmazonAvailable?: boolean;
+  isFacebookAvailable?: boolean;
+  isMicrosoftAvailable?: boolean;
+  isInstagramAvailable?: boolean;
+  isLinkedinAvailable?: boolean;
+  isGithubAvailable?: boolean;
+  isTwitterAvailable?: boolean;
+}
