+++ src/modules/SocialLogin/social-constants.ts
@@ -1,0 +1,16 @@
+export const SOCIAL_LOGIN_IDS = {
+  FACEBOOK_APP_ID: "your-facebook-app-id",
+  GOOGLE_CLIENT_ID: "your-google-client-id",
+  APPLE_CLIENT_ID: "your-apple-client-id",
+  AMAZON_CLIENT_ID: "your-amazon-client-id",
+  INSTAGRAM_APP_ID: "your-instagram-app-id",
+  INSTAGRAM_APP_SECRET: "your-instagram-app-secret",
+  MICROSOFT_APP_ID: "your-microsoft-app-id",
+  LINKEDIN_APP_ID: "your-linkedin-app-id",
+  LINKEDIN_APP_SECRET: "your-linkedin-app-secret",
+  GITHUB_APP_ID: "your-github-app-id",
+  GITHUB_APP_SECRET: "your-github-app-secret", // Add this line
+  TWITTER_V2_APP_KEY: "your-twitter-v2-app-key",
+};
+
+export const REDIRECT_URI = "http://localhost:3000/oauth/login";
