--- src/ui/atoms/Anchor/Anchor.scss
@@ -1,12 +1,0 @@
-/* Anchor Styling */
-.anchor {
-    color: #3a81cc;
-    text-decoration: none;
-    font-weight: 500;
-    transition: color 0.3s ease, text-decoration 0.3s ease;
-}
-
-.anchor:hover {
-    text-decoration: underline;
-    color: #002851;
-}
