+++ src/ui/atoms/Anchor/Anchor.tsx
@@ -1,4 +1,3 @@
-import "./Anchor.scss";
 import { AnchorProps } from "./Anchor.types";
 
 export const Anchor = ({ children, href, className, ...rest }: AnchorProps) => {
