--- src/ui/atoms/FormInput/FormInput.scss
@@ -1,21 +1,0 @@
-/* Input Styling */
-.form-input {
-    padding: 10px;
-    border: 1px solid #ccc;
-    border-radius: 8px;
-    width: 100%;
-    box-sizing: border-box;
-    font-size: 16px;
-    transition: border-color 0.3s ease, box-shadow 0.3s ease;
-}
-
-.form-input:focus {
-    border-color: #002851;
-    outline: none;
-    box-shadow: 0 0 5px rgba(0, 123, 255, 0.4);
-}
-
-.form-input:disabled{
-    background-color: #f4f4f4;
-    cursor: not-allowed;
-}
