--- src/ui/atoms/FormInput/FormInput.tsx
@@ -1,5 +1,0 @@
-import "./FormInput.scss";
-import { FormInputProps } from "./FormInput.types";
-export const FormInput = ({ className, ...rest }: FormInputProps) => {
-  return <input className={`form-input ${className || ""}`} {...rest} />;
-};
