--- src/ui/atoms/FormInput/FormInput.types.ts
@@ -1,4 +1,0 @@
-export interface FormInputProps
-  extends React.InputHTMLAttributes<HTMLInputElement> {
-  className?: string;
-}
