--- src/ui/atoms/FormInput/index.ts
@@ -1,1 +1,0 @@
-export { FormInput } from "./FormInput";
