+++ src/ui/atoms/FormLabel/FormLabel.scss
@@ -1,10 +1,11 @@
-/* Label Styling */
-.form-label {
-    font-size: 16px;
-    color: #333;
-    margin-bottom: 8px;
-    display: block;
-    font-weight: 600;
-    text-align: left;
-}
+$font-size: 14px;
+$font-weight: 500;
+$color: #000;
 
+.form-label {
+  font-size: $font-size;
+  font-weight: $font-weight;
+  color: $color;
+  display: flex;
+  justify-content: start;
+}
