+++ src/ui/atoms/FormLabel/FormLabel.tsx
@@ -3,7 +3,7 @@ import { FormLabelProps } from "./FormLable.types";
 
 export const FormLabel = ({ children, className, ...rest }: FormLabelProps) => {
   return (
-    <label className={`form-label ${className || ""}`} {...rest}>
+    <label className={`form-label ${className ?? ""}`} {...rest}>
       {children}
     </label>
   );
